requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2
ftfy==6.1.1
idna==3.4
numpy==1.24.1
opencv-python==4.7.0.68
packaging==22.0
Pillow==9.4.0
pytesseract==0.3.10
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
wcwidth==0.2.5
